fixed feedback on utils
